Fix space issue (#37)
* Addressed spaces in files not working with copying.  Fixed issue with s3 native copy not being used if neither file has options set.  Allow target ACL to be used in preference to source ACL when copying.

* Add backed intergration tests for copying with spaces

* Fixed file test to correctly expect %2520 from an literal %20

* use PathEscape instead of EscapedPath to ensure encoding literals are preserved (double encoding)

* fix test for encoded slash

+++ backend/mem/file.go
@@ -149,7 +149,7 @@ func (f *File) Seek(offset int64, whence int) (int64, error) {
 
 	length := len(f.contents)
 
-	if num := int64(length) + int64(offset) + int64(whence); num == 0 {
+	if num := int64(length) + offset + int64(whence); num == 0 {
 		return 0, nil
 	}
 	switch whence {
